+++ xapian-core/net/remoteconnection.cc
@@ -44,7 +44,7 @@ RemoteConnection::RemoteConnection(int fdin_, int fdout_,
     overlapped.hEvent = CreateEvent(NULL, FALSE, FALSE, NULL);
     if (!overlapped.hEvent)
 	throw Xapian::NetworkError("Failed to setup OVERLAPPED",
-				   context, -GetLastError());
+				   context, -(int)GetLastError());
 
 #endif
 }
@@ -87,7 +87,7 @@ RemoteConnection::read_at_least(size_t min_len, const OmTime & end_time)
 	    // Get the final result of the read.
 	    if (!GetOverlappedResult(hin, &overlapped, &received, FALSE))
 		throw Xapian::NetworkError("Failed to get overlapped result",
-					   context, -GetLastError());
+					   context, -(int)GetLastError());
 	}
 
 	if (received > 0) {
@@ -217,7 +217,7 @@ RemoteConnection::send_message(char type, const string &message, const OmTime & 
 	    // Get the final result.
 	    if (!GetOverlappedResult(hout, &overlapped, &n, FALSE))
 		throw Xapian::NetworkError("Failed to get overlapped result",
-					   context, -GetLastError());
+					   context, -(int)GetLastError());
 	}
 
 	if (n >= 0) {
